+++ front/app/containers/Admin/dashboard/users/charts/BarChartByCategory.tsx
@@ -28,7 +28,7 @@ import {
   GraphCardTitle,
   GraphCard,
   GraphCardInner,
-} from 'components/admin/Chart';
+} from 'components/admin/GraphWrappers';
 
 // resources
 import GetSerieFromStream from 'resources/GetSerieFromStream';
