+++ front/app/modules/commercial/insights/admin/containers/Reports/charts/HorizontalBarChartWithoutStream.tsx
@@ -29,7 +29,7 @@ import {
   GraphCardTitle,
   GraphCard,
   GraphCardInner,
-} from 'components/admin/Chart';
+} from 'components/admin/GraphWrappers';
 
 // types
 import { IGraphPoint } from 'typings';
