+++ front/app/modules/commercial/insights/admin/containers/Reports/charts/IdeasByStatusChart.tsx
@@ -28,7 +28,7 @@ import {
   GraphCardTitle,
   GraphCard,
   GraphCardInner,
-} from 'components/admin/Chart';
+} from 'components/admin/GraphWrappers';
 
 // resources
 import GetSerieFromStream from 'resources/GetSerieFromStream';
